Rewrite Menu
