+++ packages/bento-design-system/src/DateField/CalendarHeader.tsx
@@ -3,7 +3,7 @@ import { FunctionComponent } from "react";
 import { IconButtonProps } from "../IconButton/IconButton";
 import { IconChevronLeft, IconChevronRight } from "../Icons";
 import { Box, Column, Columns } from "../internal";
-import { MenuProps } from "../Menu/createMenu";
+import { MenuProps } from "../Menu/Menu";
 import { useDefaultMessages } from "../util/useDefaultMessages";
 import { DateFieldConfig } from "./Config";
 import { createSelector } from "./Selector";
