+++ packages/bento-design-system/src/DateField/Selector.tsx
@@ -3,7 +3,7 @@ import { useDateFormatter } from "@react-aria/i18n";
 import { FunctionComponent, useMemo } from "react";
 import { Box, Column, Columns } from "..";
 import { ListProps } from "../List/List";
-import { MenuProps } from "../Menu/createMenu";
+import { MenuProps } from "../Menu/Menu";
 import { Label } from "../Typography/Label/Label";
 import { DateFieldConfig } from "./Config";
 import { selector } from "./DateField.css";
